+++ REPTILE/__init__.py
@@ -1,6 +1,6 @@
 from REPTILE.FissionFragmentSpectrum import *
 from REPTILE.EffectiveMass import *
-from REPTILE.PowerMonitor import *
 from REPTILE.ReactionRate import *
+from REPTILE.Computables import *
 from REPTILE.SpectralIndex import *
 from REPTILE.CoverE import *
